Debug mode reverts to the uncompressed URL string, swapping readInt for readFloat on room coordinates

+++ js/model.js
@@ -257,7 +257,8 @@ function roomToString(room) {
 function roomFromString(string) {
     var s = string.split(",");
     var room = new Room(getRoomMetadata(s[0]));
-    room.setPosition(parseInt(s[1]), parseInt(s[2]), parseInt(s[3]), parseInt(s[4]) * 90);
+    // room coordinates may be fractional because of old dojo rooms, floor and rotation are still ints
+    room.setPosition(parseFloat(s[1]), parseFloat(s[2]), parseInt(s[3]), parseInt(s[4]) * 90);
     return room;
 }
